Move input payload helpers into package
