+++ src/heart/peripheral/radio.py
@@ -10,7 +10,7 @@ from dataclasses import dataclass
 from typing import Any, Iterator, Mapping
 
 from heart.peripheral.core import Peripheral
-from heart.peripheral.input_payloads import RadioPacket
+from heart.peripheral.input_payloads.radio import RadioPacket
 from heart.utilities.logging import get_logger
 
 serial: Any
